requirements.txt
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
secure-smtplib
jinja2
Flask
APScheduler
gunicorn
google-auth